Fix PC build errors and runtime crashes.
Build fixes:
- CMakeLists.txt: define M_PI_F / M_PI_2_F (ARM GCC extensions, absent on UCRT)
- lib/vector.h: add static to inline rand_between() — non-static inline in a
  header has no external definition in C11, causing linker errors on x86-64
- CMakeLists.txt: remove SDL2main linkage (not needed with SDL_MAIN_HANDLED,
  and caused undefined reference to SDL_strlen etc. due to link order)

Runtime fixes:
- platform/pc/main_pc.c: switch from SDL_Surface/BlitScaled to SDL_Renderer +
  SDL_Texture with manual 8bpp→ARGB conversion each frame; SDL_BlitScaled is
  unreliable with palettized surfaces across SDL2 versions
- lib/debug.h/c, src/flow-field.c, src/sequence.c: change key_callback params
  and all callback functions from u32 to uintptr_t so variable addresses are
  not truncated on 64-bit PC (was causing segfault on G, 1-9 etc.)

+++ lib/debug.c
@@ -14,8 +14,8 @@ INCBIN_EXTERN(debug_font);
 struct debug_key {
     u8 key_code;
     key_callback key_func;
-    u32 key_param1;
-    u32 key_param2;
+    uintptr_t key_param1;
+    uintptr_t key_param2;
 };
 
 static u8 debug_font_mode13[Debug_MaxGlyphs * 8 * 8];
@@ -93,7 +93,7 @@ void debug_write_vidc(u32 vidc_reg)
 #endif
 }
 
-void debug_register_key(u8 key_code, key_callback key_func, u32 key_param1, u32 key_param2)
+void debug_register_key(u8 key_code, key_callback key_func, uintptr_t key_param1, uintptr_t key_param2)
 {
     if (debug_num_keys < Debug_MaxKeys) {
         debug_key_list[debug_num_keys].key_code = key_code;
@@ -135,18 +135,18 @@ void debug_do_keypress_callbacks()
     }
 }
 
-void debug_toggle_word(u32 addr, u32 val)
+void debug_toggle_word(uintptr_t addr, uintptr_t val)
 {
     (void)val;
     *((u32*)addr) ^= 1;
 }
 
-void debug_set_word(u32 addr, u32 val)
+void debug_set_word(uintptr_t addr, uintptr_t val)
 {
-    *((u32*)addr) = val;
+    *((u32*)addr) = (u32)val;
 }
 
-void debug_word_add(u32 addr, u32 val)
+void debug_word_add(uintptr_t addr, uintptr_t val)
 {
     *((int*)addr) += (int)val;
 }
